refactored vesselInfo into own namespace
