+++ app/src/actions/heatmap.js
@@ -10,7 +10,7 @@ import {
 } from 'util/heatmapTileData';
 import { LOADERS } from 'config';
 import { LAYER_TYPES } from 'constants';
-import { clearVesselInfo, addVessel, hideVesselsInfoPanel } from 'actions/vesselInfo';
+import { clearVesselInfo, addVessel, hideVesselsInfoPanel } from 'vesselInfo/vesselInfoActions';
 import { setEncountersInfo, clearEncountersInfo } from 'encounters/encountersActions';
 import { trackMapClicked } from 'analytics/analyticsActions';
 import { addLoader, removeLoader, zoomIntoVesselCenter } from 'actions/map';
