+++ app/src/actions/map.js
@@ -1,4 +1,4 @@
-import { clearVesselInfo } from 'actions/vesselInfo';
+import { clearVesselInfo } from 'vesselInfo/vesselInfoActions';
 import { trackCenterTile } from 'analytics/analyticsActions';
 import { ANALYTICS_TILE_COORDS_SCALE, ANALYTICS_TRACK_DRAG_FROM_ZOOM, CLUSTER_CLICK_ZOOM_INCREMENT } from 'config';
