+++ app/src/analytics/analyticsMiddleware.js
@@ -26,7 +26,7 @@ import isFunction from 'lodash/isFunction';
 import { SEARCH_QUERY_MINIMUM_LIMIT, TIMELINE_SPEED_CHANGE } from 'config';
 
 import { FLAGS } from 'app/src/constants';
-import { TOGGLE_VESSEL_PIN, SET_PINNED_VESSEL_HUE } from 'actions/vesselInfo';
+import { TOGGLE_VESSEL_PIN, SET_PINNED_VESSEL_HUE } from 'vesselInfo/vesselInfoActions';
 import { SET_WORKSPACE_ID } from 'workspace/workspaceActions';
 
 const GA_ACTION_WHITELIST = [
