+++ app/src/encounters/containers/EncountersVessel.js
@@ -1,7 +1,7 @@
 import { connect } from 'react-redux';
 import EncountersVessel from 'encounters/components/EncountersVessel';
 import { login } from 'user/userActions';
-import { addVessel, clearVesselInfo } from 'actions/vesselInfo';
+import { addVessel, clearVesselInfo } from 'vesselInfo/vesselInfoActions';
 import { clearEncountersInfo } from 'encounters/encountersActions';
 
 const mapStateToProps = state => ({
