+++ app/src/recentVessels/containers/RecentVesselItem.js
@@ -1,6 +1,6 @@
 import { connect } from 'react-redux';
 import RecentVesselItem from 'recentVessels/components/RecentVesselItem';
-import { addVessel, clearVesselInfo } from 'actions/vesselInfo';
+import { addVessel, clearVesselInfo } from 'vesselInfo/vesselInfoActions';
 import { toggleLayerVisibility } from 'layers/layersActions';
 import { setRecentVesselsModalVisibility } from 'recentVessels/recentVesselsActions';
 import { trackRecentVesselAdded } from 'analytics/analyticsActions';
